Opt 9 Clinton

+++ mjpeg_encoder.h
@@ -23,9 +23,9 @@
 
 struct yuv
 {
-  uint8_t *Y;
-  uint8_t *U;
-  uint8_t *V;
+  float *Y;
+  float *U;
+  float *V;
 };
 
 struct dct
